+++ app/services/reverse_geocoding/places/fetch_data.rb
@@ -19,7 +19,6 @@ class ReverseGeocoding::Places::FetchData
 
     first_place = reverse_geocoded_places.shift
     update_place(first_place)
-    add_suggested_place_to_a_visit
     reverse_geocoded_places.each { |reverse_geocoded_place| fetch_and_create_place(reverse_geocoded_place) }
   end
 
@@ -52,22 +51,10 @@ class ReverseGeocoding::Places::FetchData
     new_place.source = :photon
 
     new_place.save!
-
-    add_suggested_place_to_a_visit(suggested_place: new_place)
   end
 
   def reverse_geocoded?
     place.geodata.present?
-  end
-
-  def add_suggested_place_to_a_visit(suggested_place: place)
-    visits = Place.near([suggested_place.latitude, suggested_place.longitude], 0.1).flat_map(&:visits)
-
-    visits.each do |visit|
-      next if visit.suggested_places.include?(suggested_place)
-
-      visit.suggested_places << suggested_place
-    end
   end
 
   def find_place(place_data)